add&mod: 将SimulateAction改为扩展方法，不再继承PostMessageSimulator和InputSimulatorExtension；对于部分不支持的按键和功能（如PostMessageSimulator中的鼠标侧键）予以忽略处理
